feat: move ppattern to use internally
