+++ api/src/main/java/jakarta/data/metamodel/Pattern.java
@@ -15,11 +15,9 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.repository;
+package jakarta.data.metamodel;
 
 import jakarta.data.Operator;
-import jakarta.data.metamodel.Restriction;
-
 
 
 /**
@@ -41,7 +39,7 @@ import jakarta.data.metamodel.Restriction;
  *
  * @param <T> the type of the entity on which the restriction is applied.
  */
-public record Pattern<T>(String field, String value, boolean ignoreCase) implements Restriction<T> {
+record Pattern<T>(String field, String value, boolean ignoreCase) implements Restriction<T> {
 
     /**
      * The operator for pattern-based restrictions, fixed as `LIKE`.
