+++ api/src/main/java/jakarta/data/metamodel/TextAttribute.java
@@ -18,7 +18,6 @@
 package jakarta.data.metamodel;
 
 import jakarta.data.Sort;
-import jakarta.data.repository.Pattern;
 
 
 /**
